requirements.txt
alembic==0.7.7
facebook-sdk==0.4.0
Flask==0.10.1
Flask-Migrate==1.4.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-WTF==0.12
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.1
MarkupSafe==0.23
psycopg2==2.6.1
SQLAlchemy==1.0.8
Werkzeug==0.10.4
WTForms==2.0.2
